requirements.txt
aiohttp
aiosqlite
backoff
discord.py[voice]
python-dotenv
humanize
matplotlib
motor
psutil
pyfiglet
pymongo
pytz
PyNaCl
deep-translator>=1.11.4
numpy>=1.20.0
aiofiles>=0.8.0
Pillow>=10.0.0
python-snappy>=0.6.0
hypothesis>=6.0.0
pytest>=7.0.0
